project/requirements.txt
mysql-connector-python==8.0.18
python-dotenv==0.10.3